% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acids.R
\name{amino_acids}
\alias{amino_acids}
\title{The 20 standard amino acids}
\usage{
amino_acids()
}
\value{
Three-letter codes of the standard amino acids.
}
\description{
The 20 amino acids that are encoded directly by the codons of the universal
genetic code.
}
\examples{
amino_acids()

}
