% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acids_properties.R
\docType{data}
\name{amino_acids_properties}
\alias{amino_acids_properties}
\title{Amino acid side chain property values}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 4 columns.
}
\source{
Table 1,  Grantham (1974), \doi{10.1126/science.185.4154.862}.
}
\usage{
amino_acids_properties
}
\description{
A dataset containing the amino acid side chain property values
---composition, polarity and molecular volume. These values were obtained
from Table 1,  Grantham (1974), \doi{10.1126/science.185.4154.862}.
}
\examples{
amino_acids_properties

}
\keyword{datasets}
