% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grantham_distance.R
\name{grantham_distance_original}
\alias{grantham_distance_original}
\title{Grantham's distance (original)}
\source{
\doi{10.1126/science.185.4154.862}.
}
\usage{
grantham_distance_original(x, y)
}
\arguments{
\item{x}{A character vector of amino acid three-letter codes.}

\item{y}{A character vector of amino acid three-letter codes.}
}
\value{
A \link[tibble:tibble-package]{tibble} of Grantham's distances for each
amino acid pair.
}
\description{
This function calculates the Grantham's distance for pairs of amino acids. It
uses the pre-calculated distances for each amino acid pair as published in
Table 2 of Science (1974). 185(4154): 862--4 by R. Grantham.
}
\keyword{internal}
