% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grantham_distances_matrix.R
\docType{data}
\name{grantham_distances_matrix}
\alias{grantham_distances_matrix}
\title{Grantham distance matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20 rows and 20 columns.
}
\source{
Table 2,  Grantham (1974), \doi{10.1126/science.185.4154.862}.
}
\usage{
grantham_distances_matrix
}
\description{
A dataset containing Grantham distances in the format of a matrix. These
values were obtained from Table 2,  Grantham (1974),
\doi{10.1126/science.185.4154.862}.
}
\examples{
grantham_distances_matrix

}
\keyword{datasets}
