% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ij2k.R
\name{ij2k}
\alias{ij2k}
\title{Convert an (i, j) index to a linear index.}
\usage{
ij2k(i, j, n)
}
\arguments{
\item{i}{i index, i.e. row position; indexing starts at 1.}

\item{j}{j index, i.e. column position; indexing starts at 1.}

\item{n}{size of the square matrix.}
}
\value{
Linear position.
}
\description{
Converts an (i, j) index to a linear index. Converts the double index of
a square matrix to the corresponding linear one. This is column-major
as it is default in R.
}
\keyword{internal}
